Some changes include

PostControl has new API Calls
PostControlTest File was created in the Test Folder
Added React code for sending the Post to the database in Post.Create.tsx

+++ SavorySpringAPI/src/main/java/com/savory/savoryAPI/post/dto/PostsDto.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Builder
 @Value
 public class PostsDto {
-    long post_id;
-    long userID;
+    int post_id;
+    int userID;
 //    Person person;
     String headline;
     String ingredients;
